+++ src/InfosUsuario.js
@@ -1,0 +1,13 @@
+import React from "react"
+
+
+
+export default function InfoUsuario() {
+    const titulo = "Marcelo Bigaram"
+    return (
+        <>
+        <img src="https://picsum.photos/400/400?a=2 " alt="" />
+      <h4>{titulo}</h4>
+        </>
+    );
+}
