+++ src/styles.css
@@ -39,9 +39,12 @@ main {
   display: grid;
   grid-template-columns: repeat(4, 1fr);
   grid-template-rows: repeat(4, 1fr);
-  row-gap: 10px;
+  row-gap: 5px;
+  column-gap: 5px;
+  margin: 5px;
+  /* row-gap: 10px;
   column-gap: 10px;
-  margin: 10px;
+  margin: 10px; */
 }
 
 .box-pagina-principal {
@@ -66,7 +69,8 @@ footer {
 
 .box-pagina-principal > img {
   width: 100%;
-  height: 200px;
+  height: 150px;
+  /* height: 200px; */
 }
 
 h4 {
